+++ lib/baby_squeel/active_record.rb
@@ -17,6 +17,16 @@ module BabySqueel
       def ordering(&block)
         order DSL.evaluate(self, &block)
       end
+
+      # Constructs Arel for ActiveRecord::Base#group using the DSL.
+      def grouping(&block)
+        group DSL.evaluate(self, &block)
+      end
+
+      # Constructs Arel for ActiveRecord::Base#having using the DSL.
+      def having_grouped(&block)
+        having DSL.evaluate(self, &block)
+      end
     end
 
     module WhereChain